Combine image options types
